+++ linmod/models.py
@@ -1,7 +1,14 @@
+import string
+from itertools import product
+from typing import Iterable
+
 import numpy as np
 import numpyro
 import numpyro.distributions as dist
 import polars as pl
+from numpyro.diagnostics import summary
+from numpyro.infer.mcmc import MCMC
+from plotnine import aes, geom_line, ggplot, theme_bw
 
 from .utils import expand_grid, pl_softmax
 
@@ -418,3 +425,127 @@ def multinomial_likelihood(
     assert z.shape == (N.shape[0], beta_0.shape[1])
 
     return dist.Multinomial(total_count=N, logits=z)
+
+
+def get_convergence(
+    mcmc: MCMC, ignore_nan_in: list[str] = [], drop_ignorable_nan: bool = True
+) -> pl.DataFrame:
+    allowed = set(string.ascii_lowercase + string.ascii_uppercase + "_")
+    for ignore in ignore_nan_in:
+        assert set(ignore).issubset(allowed)
+
+    raw_summary = summary(mcmc.get_samples(group_by_chain=True))
+
+    convergence = pl.DataFrame(
+        {
+            "param": [],
+            "param_no_dim": [],
+            "n_eff": [],
+            "r_hat": [],
+        }
+    ).cast(
+        {
+            "param": pl.String,
+            "param_no_dim": pl.String,
+            "n_eff": pl.Float64,
+            "r_hat": pl.Float64,
+        }
+    )
+
+    for k, v in raw_summary.items():
+        names_nodim = []
+        names = []
+        ess = []
+        psrf = []
+
+        if len(v["n_eff"].shape) == 0:
+            names = [k]
+            names_nodim = [k]
+            ess = [float(v["n_eff"])]
+            psrf = [float(v["r_hat"])]
+        else:
+            array_indices = [
+                idx for idx in product(*map(range, v["n_eff"].shape))
+            ]
+            names = [
+                k + "[" + ",".join([str(i) for i in idx]) + "]"
+                for idx in array_indices
+            ]
+            names_nodim = [k] * len(array_indices)
+            ess = [float(v["n_eff"][idx]) for idx in array_indices]
+            psrf = [float(v["r_hat"][idx]) for idx in array_indices]
+
+        param = pl.DataFrame(
+            {
+                "param": names,
+                "param_no_dim": names_nodim,
+                "n_eff": ess,
+                "r_hat": psrf,
+            }
+        )
+
+        convergence = pl.concat([convergence, param])
+
+    nans = convergence.filter(
+        (pl.col("n_eff").is_nan()) | (pl.col("r_hat").is_nan())
+    )
+    if nans.shape[0] > 0:
+        nans = nans.filter(~pl.col("param_no_dim").is_in(ignore_nan_in))
+        if nans.shape[0] > 0:
+            bad = nans["param"].unique().to_list()
+            raise RuntimeError(
+                "Found unexpected NaN convergence values for parameters: "
+                + str(bad)
+            )
+        if drop_ignorable_nan:
+            convergence = convergence.filter(
+                (pl.col("n_eff").is_not_nan()) & (pl.col("r_hat").is_not_nan())
+            )
+
+    return convergence.drop("param_no_dim")
+
+
+def plot_convergence(mcmc: MCMC, params: Iterable[str]):
+    posterior = mcmc.get_samples(group_by_chain=True)
+    plots = []
+    for par in params:
+        if par.count("[") > 0:
+            dimless_par = par.split("[")[0]
+            indices = [int(i) for i in par.split("[")[1][:-1].split(",")]
+            index_tuple = tuple(
+                [*[slice(None), slice(None)], *[i for i in indices]]
+            )
+            chains = posterior[dimless_par][index_tuple]
+        else:
+            chains = posterior[par]
+
+        niter = chains.shape[1]
+        df = pl.concat(
+            [
+                pl.DataFrame(
+                    {
+                        "param": [par] * niter,
+                        "chain": [str(chain)] * niter,
+                        "iteration": list(range(niter)),
+                        "value": chains[chain].tolist(),
+                    }
+                )
+                for chain in range(chains.shape[0])
+            ]
+        )
+
+        plt = (
+            ggplot(df)
+            + geom_line(
+                aes(
+                    x="iteration",
+                    y="value",
+                    color="chain",
+                ),
+            )
+            + theme_bw(base_size=20)
+        )
+
+        plots.append(plt)
+
+    return plots
